+++ 8LETTE-TextRPG/ItemFolder/EquipmentType.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace _8LETTE_TextRPG
+namespace _8LETTE_TextRPG.ItemFolder
 {
     public enum EquipmentType
     {
