+++ 8LETTE-TextRPG/ItemFolder/Item.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text;
 
-namespace _8LETTE_TextRPG
+namespace _8LETTE_TextRPG.ItemFolder
 {
     public class Item : IEquipable, IUsable
     {
