+++ 8LETTE-TextRPG/MonsterFolder/Monster.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace _8LETTE_TextRPG
+namespace _8LETTE_TextRPG.MonsterFolder
 {
     public abstract class Monster
     {
