+++ 8LETTE-TextRPG/MonsterFolder/MonsterType.cs
@@ -4,10 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace _8LETTE_TextRPG
+namespace _8LETTE_TextRPG.MonsterFolder
 {
-    public interface IUsable
+    public enum MonsterType
     {
-        void Use();
+        InfLoop,
+        // 여기에 몬스터 종류 추가
     }
 }
