+++ 8LETTE-TextRPG/MonsterSpawner.cs
@@ -1,4 +1,6 @@
-﻿namespace _8LETTE_TextRPG
+﻿using _8LETTE_TextRPG.MonsterFolder;
+
+namespace _8LETTE_TextRPG
 {
     internal class MonsterSpawner
     {
