+++ 8LETTE-TextRPG/Quest.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using _8LETTE_TextRPG.ItemFolder;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
