+++ 8LETTE-TextRPG/ScreenFolder/InventoryScreen.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace _8LETTE_TextRPG
+namespace _8LETTE_TextRPG.ScreenFolder
 {
     internal class InventoryScreen : Screen
     {
