+++ 8LETTE-TextRPG/ScreenFolder/ShopScreen.cs
@@ -1,6 +1,7 @@
-﻿using static _8LETTE_TextRPG.Monster;
+﻿using _8LETTE_TextRPG.ItemFolder;
+using static _8LETTE_TextRPG.MonsterFolder.Monster;
 
-namespace _8LETTE_TextRPG
+namespace _8LETTE_TextRPG.ScreenFolder
 {
     internal class ShopScreen : Screen
     {
