+++ 8LETTE-TextRPG/ScreenFolder/UseItemScreen.cs
@@ -1,8 +1,10 @@
-﻿namespace _8LETTE_TextRPG
+﻿using _8LETTE_TextRPG.ItemFolder;
+
+namespace _8LETTE_TextRPG.ScreenFolder
 {
-    internal class EquipmentScreen : Screen
+    internal class UseItemScreen : Screen
     {
-        public static readonly EquipmentScreen Instance = new EquipmentScreen();
+        public static readonly UseItemScreen Instance = new UseItemScreen();
 
         private Item? _selectedItem = null;
 
@@ -10,18 +12,18 @@
         {
             Console.Clear();
 
-            PrintTitle("인벤토리 - 장착 관리");
+            PrintTitle("인벤토리 - 아이템 사용");
 
-            Console.WriteLine("보유 중인 아이템을 장착할 수 있습니다.");
+            Console.WriteLine("보유 중인 아이템을 사용할 수 있습니다.");
             Console.WriteLine();
 
             Console.WriteLine("[아이템 목록]");
-            Player.Instance.Inventory.ShowPlayerItems(Inventory.PrintState.Equipment);
+            Player.Instance.Inventory.ShowPlayerItems(Inventory.PrintState.Usable);
             Console.WriteLine();
 
             if (_selectedItem != null)
             {
-                Console.WriteLine($"{_selectedItem.Name}을(를) {(_selectedItem.IsEquipped ? "장착" : "해제")}했습니다.");
+                Console.WriteLine($"{_selectedItem.Name}을(를) 사용했습니다.");
                 Console.WriteLine();
             }
 
@@ -35,32 +37,22 @@
             string? input = Console.ReadLine();
             if (input == "0")
             {
-                _selectedItem = null;
                 return InventoryScreen.Instance;
             }
             else if (int.TryParse(input, out int num))
             {
-                Item[] items = Player.Instance.Inventory.GetAllItems(ItemType.Equipment);
+                Item[] items = Player.Instance.Inventory.GetAllItems(ItemType.Usable);
                 if (num < 1 || num > items.Length)
                 {
-                    _selectedItem = null;
                     _isRetry = true;
                     return this;
                 }
 
                 _selectedItem = items[num - 1];
-                if (!_selectedItem.IsEquipped)
-                {
-                    Player.Instance.Inventory.Equip(_selectedItem);
-                }
-                else
-                {
-                    Player.Instance.Inventory.Unequip(_selectedItem);
-                }
+                _selectedItem.Use();
             }
             else
             {
-                _selectedItem = null;
                 _isRetry = true;
             }
